dp14: Adding MST-Lite Audio Enable/Disable support
This patch adds the support to enable/disable Audio in
dp14 txss and rxss based MST system


Acked by: Kapil Usgaonkar <kapilu@xilinx.com>

+++ XilinxProcessorIPLib/drivers/dp14/src/xdp_mst.c
@@ -2472,7 +2472,7 @@ XDp_RxIicMapEntry *XDp_RxGetIicMapEntry(XDp *InstancePtr, u8 PortNum,
  *
 *******************************************************************************/
 u32 XDp_RxSetIicMapEntry(XDp *InstancePtr, u8 PortNum, u8 IicAddress,
-						u8 ReadNumBytes, u8 *ReadData)
+						u16 ReadNumBytes, u8 *ReadData)
 {
 	XDp_RxIicMapEntry *IicMapEntry;
